Criada página de cadastro inicial (e-mail, senha, confirmação de senha)
