+++ src/pages/Signup/styles.css
@@ -1,21 +1,98 @@
-.full-cadastro {
-  background-color: var(--color-background);
-  min-height: calc(100vh - 72px);
-  position: relative;
+.corpo{
+    display: flex;
+    flex-direction: column;
+    width: 100vw;
+    background-color: var(--color-background);
+    height: 100vh;
 }
 
-.form-box {
-  border-radius: 15px;
+.informacoes{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    font-size: 0.9em;
+    text-align: center;
+    margin-top: 1em;
+    color: var(--color-text-title);
 }
 
-.btn-cadastro {
-  display: inline-block;
-  margin-left: auto;
-  border-radius: 25px;
+
+.btnCadastro{
+    background-color: var(--color-text-title);
+    color: white;
+    font-weight: 700;
+    text-align: center;
+    width: 10em;
+    height: 4em;
+    padding: 0.5em;
+    border-radius: 2em;
 }
 
-#cadastro-footer {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
+.formularioLogin{
+    display: flex;
+    flex-direction: column;
+    margin-top: 1em;
+    background-color: var(--color-text-base);
+    align-items: center;
+    color: white;
+    font-weight: 700;
+    border-radius: 1em;
+    padding: 1em;
+}
+
+.btnRec{
+    color: white;
+}
+
+
+.inputLogin{
+ width: 85vw;
+ height: 2em;
+ border-radius: 0.5em;
+ border: none;
+}
+
+.opcoesLogin{
+    display: flex;
+}
+
+
+.btnLogin{
+    margin-right: 2em;
+    margin-bottom: 1em;
+    color: white;
+    background-color: var(--color-text-title);
+    font-weight: 700;
+    width: 7em;
+    height: 2.1em;
+    border-radius: 1em;
+    border: 1px var(--color-text-title);
+}
+
+@media(min-width: 600px){
+    .corpo{
+        flex-direction: row;
+        justify-content: space-around;
+    }
+
+    .informacoes{
+        width: 40vw;
+    }
+    .formularioLogin{
+      width: 40vw;
+      height: 40vh;
+    }
+
+    .inputLogin{
+        width: 35vw; 
+        height: 6vh
+       }
+
+    .opcoesLogin{
+        justify-content: space-between;
+        width: 89%;
+        margin-top: 1.5em;
+    }
+
+    
 }
